backend/requirements.txt
tableauserverclient==0.38
python-dotenv==1.1.1
pandas==2.3.1
openpyxl==3.1.5
tqdm==4.67.1